// spring-ai-alibaba-studio/spring-ai-alibaba-studio-server/frontend/packages/spark-flow/src/demos/flow-demo/types/flow.ts
export interface IVariable extends Omit<INodeDataInputParamItem, 'key'> {
  id: string;
}

export interface IVariableHandleGroupItem {
  group_id: string;
  group_name: string;
  output_type: IValueType;
  variables: IVariable[];
}

export interface IVariableHandleNodeParam {
  type: 'group' | 'json' | 'template';
  /* When type is not group, group_strategy and groups need to be filled in */
  group_strategy: 'firstNotNull' | 'lastNotNull';
  groups: IVariableHandleGroupItem[];
  /* When type is json, json_params need to be filled in */
  json_params: INodeDataInputParamItem[];
  /* When type is template, template_content needs to be filled in */
  template_content: string;
}
// ...
export type INodeDataNodeParam =
  | IStartNodeParam
  | IInputNodeParam
  | IEndNodeParam
  | IOutputNodeParam
  | ILLMNodeParam
  | IParameterExtractorNodeParam
  | IJudgeNodeParam
  | IApiNodeParam
  | IParallelNodeParam
  | IIteratorNodeParam
  | IAppComponentNodeParam
  | IPluginNodeParam
  | IClassifierNodeParam
  | IScriptNodeParam
  | IVariableAssignNodeParam
  | IVariableHandleNodeParam
  | IRetrievalNodeParam
  | IMCPNodeParam;

export type IWorkFlowNodeData<T = INodeDataNodeParam> = {
  /* Node name */
  label: string;
  /* Node input parameters */
  input_params: INodeDataInputParamItem[];
  /* Node output parameters */
  output_params: INodeDataOutputParamItem[];
  /* Node business parameters */
  node_param: T;
  /* Node description */
  desc?: string;
};
// ...
export type IVariableHandleNodeData =
  IWorkFlowNodeData<IVariableHandleNodeParam>;